+++ layout/drawers.php
@@ -24,17 +24,13 @@
  * * Include activity navigation
  * * Include course related hints
  * * Include back to top button
-<<<<<<< HEAD
  * * Include scroll spy
  * * Include footnote
  * * Include static pages
  * * Include Jvascript disabled hint
  * * Include advertisement tiles
  * * Include info banners
-=======
- * * Include activity navigation
  * * Include additional block regions
->>>>>>> 952a601 (Implemented additional block regions)
  *
  * @package   theme_boost_union
  * @copyright 2022 Luca Bösch, BFH Bern University of Applied Sciences luca.boesch@bfh.ch
@@ -133,11 +129,7 @@ $templatecontext = [
     'hasregionmainsettingsmenu' => !empty($regionmainsettingsmenu),
     'overflow' => $overflow,
     'headercontent' => $headercontent,
-    'addblockbutton' => $addblockbutton,
-    'user_is_editing' => $OUTPUT->page->user_is_editing(),
-    'outsideregionsplacement' => get_config('theme_boost_union', 'outsideregionsplacement') ==
-            THEME_BOOST_UNION_SETTING_OUTSIDEREGIONSPLACEMENT_NEXTMAINCONTENT ?
-                    'blocks-next-maincontent' : 'blocks-near-windowedges',
+    'addblockbutton' => $addblockbutton
 ];
 
 // Include the template content for the course related hints.