+++ locallib.php
@@ -1272,14 +1272,14 @@ function theme_boost_union_get_course_header_image_url() {
 
         // Get all files from filearea.
         $files = $fs->get_area_files($systemcontext->id, 'theme_boost_union', 'courseheaderimagefallback',
-                false, 'itemid', false);
+            false, 'itemid', false);
 
         // Just pick the first file - we are sure that there is just one file.
         $file = reset($files);
 
         // Build and return the image URL.
         return moodle_url::make_pluginfile_url($file->get_contextid(), $file->get_component(), $file->get_filearea(),
-                $file->get_itemid(), $file->get_filepath(), $file->get_filename());
+            $file->get_itemid(), $file->get_filepath(), $file->get_filename());
     }
 
     // As no picture was found, return null.
@@ -1322,17 +1322,19 @@ function theme_boost_union_set_mobilecss_url() {
  */
 function theme_boost_union_get_additional_regions($pageregions=[]) {
     $regions = [
-            'top' => 'outside-top',
             'footerleft' => 'footer-left',
             'footerright' => 'footer-right',
             'footercenter' => 'footer-center',
             'offcanvasleft' => 'offcanvas-left',
             'offcanvasright' => 'offcanvas-right',
             'offcanvascenter' => 'offcanvas-center',
-            'left' => 'outside-left',
-            'right' => 'outside-right',
-            'bottom' => 'outside-bottom',
-            'headertop' => 'header-top'
+            'outsideleft' => 'outside-left',
+            'outsideright' => 'outside-right',
+            'outsidetop' => 'outside-top',
+            'outsidebottom' => 'outside-bottom',
+            'contentupper' => 'content-upper',
+            'contentlower' => 'content-lower',
+            'header' => 'header'
     ];
 
     return ($pageregions) ? array_intersect($regions, $pageregions) : $regions;
@@ -1346,9 +1348,15 @@ function theme_boost_union_get_additional_regions($pageregions=[]) {
  */
 function theme_boost_union_get_block_regions($layout) {
 
-    $regionsettings = get_config('theme_boost_union', $layout.'regions');
+    // Get the admin setting for the layout.
+    $regionsettings = get_config('theme_boost_union', 'blockregionsfor'.$layout);
+
+    // Explode the admin setting to get the block regions.
     $settings = !empty($regionsettings) ? explode(',', $regionsettings) : [];
-    // Used the side-pre as default region for layouts.
+
+    // Add the configured regions to the side-pre region (which is always provided by Boost core).
     $regions = array_merge(['side-pre'], $settings);
+
+    // Return.
     return $regions;
 }