+++ CHANGES.md
@@ -6,7 +6,8 @@ Changes
 
 ### Unreleased
 
-* 2023-03-22 - Feature: Eleven additional block regions, solves #30
+* 2023-03-22 - Feature: Allow admin to provide several additional block regions, solves #30.
+               Please note: This is a comparably large addition. If you encounter any issues with this feature, please report it on https://github.com/moodle-an-hochschulen/moodle-theme_boost_union/issues.
 * 2023-03-20 - Improvement: Don't force child themes to reimplement various color settings (e.g. 'brandcolor'), solves #260.
 * 2023-03-17 - Improvement: Reduce code duplication when child theming by checking theme ancestry in theme_boost_union_before_standard_html_head, solves #245.